외래키 충돌

+++ MvcBoardApp/MvcBoardApp/ViewModels/CommentCounter.cs
@@ -14,7 +14,7 @@ namespace MvcBoardApp.ViewModels
 
         public void GetCount(Board board)
         {
-            board.CommentCount = CommentCount;
+            board.CommentCount = (int)CommentCount;
         }
     }
 }
